Fix broken history tests
The test code is currently setting a limit of 10 history data values.
However, the limit handling is not implemented in both history
implementations and therefore the limit is never enforced.
Removed the limit from the history tests to ensure that no tests are broken
when the limit handling is actually implemented.