Reap Client-KeepAlive Threads for reconnect
In some situations, Clients seem to create multiple KeepAlive-Threads.
The old ones are never properly stopped, and there are no references to
them to be able to stop them. This causes programs not to quit properly
or completely.

To Redeem, we now check for an existing KeepAlive-Thread whenever
creating a new one — and try to stop the existing one beforehand.