Parsing of custom structures
Previously custom structures that contain enums wouldn't be parsed correctly when calling load_type_definitions(). This is now fixed by also creating enum classes that represent the enums that are exposed in the server's address space and changing get_default_value() so that it is able to create default values for enums.