Fix: Do not quit uals when variable cannot be read
Some OPC UA servers (e.g. in B&R PLCs) have OPC UA nodes which
should be of type Object but are of type Variable. They do not
provide a value, which causes uals to fail.

This commit catches this exception and allows to continue
listing the OPC UA nodes.