[subscriptions] Fix race condition in client when creating subs
Due to how Futures are used, we could timeout on the client side, but
still sucessfully create a subscription and add callbacks in the
_receiver thread. Once this happens, receiver thread will deadlock on
receiving the publish response from the server due to a while loop that
waits for the subscription_id to be set. This will never happen as it
was supposed to be set in the main thread, which timed out and failed.

The fix is to make sure we set the subscription_id in the _receiver
thread as a separate callback. This is not the ideal fix, which would
probably include re-designing the Subscription class API in a way to
allow checking for the staus, and changing the way
Client.create_subscription works. Since we don't want to make large
changes to this - let's just go with a simple fix that removes the
deadlock, and keeps the API pretty much compatible.

This commit also contains unittests that excercise the offending code
path and clearly avoids a deadlock.