Bugfix: Node.set_value() does not work with enumerations
If a client tries to set a enum variable without explicitly force the data type it becomes implicit converted to a ua.Variant.Int64. The server expects an Int32 for an enum value. This is why set_value() crashes if you try to set an enum value using a integer literal like "set_value(2)".

The enumeration seems to need a special treadment at this point. See ua_utils.py def data_type_to_variant_type().

If no data type was given in method set_value(), this change checks the data type of the node object and automatically converts the value argument to the correct target data type of the server.