Properly shut down ThreadLoop
Stopping the InternalServer object should be a synchrounous call (once
the call returns the caller can be sure that the server if fully
stopped).

Hence, instead of just pushing a stop callback into the ThreadLoop wait
for the ThreadLoop to finish before returning.

Also re-order the shutdown to match the setup code.